clear up sm2 algorithm
